requirements.txt
beautifulsoup4
spotipy~=2.23.0
requests
python-dotenv
ytmusicapi